+++ newIDE/app/src/ObjectEditor/Editors/SpriteEditor/SpritesList.js
@@ -393,12 +393,18 @@ const SpritesList = ({
         multiSelection: true,
         resourceKind: 'image',
       });
-      resources.forEach(resource => {
-        applyResourceDefaults(project, resource);
-        project.getResourcesManager().addResource(resource);
-      });
+      if (resourceSource.shouldCreateResource) {
+        resources.forEach(resource => {
+          applyResourceDefaults(project, resource);
+          project.getResourcesManager().addResource(resource);
+        });
+      }
 
-      if (directionSpritesCountBeforeAdding === 0 && resources.length > 1) {
+      if (
+        directionSpritesCountBeforeAdding === 0 &&
+        resources.length > 1 &&
+        resourceSource.shouldGuessAnimationsFromName
+      ) {
         const resourcesByAnimation = groupResourcesByAnimations(resources);
         if (resourcesByAnimation.size > 1) {
           addAnimations(resourcesByAnimation);
@@ -416,9 +422,11 @@ const SpritesList = ({
         }
       }
 
-      // Important, we are responsible for deleting the resources that were given to us.
-      // Otherwise we have a memory leak, as calling addResource is making a copy of the resource.
-      resources.forEach(resource => resource.delete());
+      if (resourceSource.shouldCreateResource) {
+        // Important, we are responsible for deleting the resources that were given to us.
+        // Otherwise we have a memory leak, as calling addResource is making a copy of the resource.
+        resources.forEach(resource => resource.delete());
+      }
 
       forceUpdate();
